refactor(splunk): keep lifecycle state in memory
